Datasets/requirements.txt
# # cma==2.6.0
# # moderngl==5.4.2
# # numpy==1.15.4
# # objloader==0.2.0
# # Pillow==6.2.0
# # pyrr==0.9.2
# # scipy==1.2.1
# # torch==0.4.1
# # torchvision==0.2.1


# requests
# gdown
# scikit-image
# click
# dataclasses
# PySocks
# pandas
# seaborn
# ftfy
# regex
# tqdm
# pytorch_pretrained_vit

moderngl>=5.4.2
numpy>=1.15.4
objloader>=0.2.0
opencv-python>=3.4.3
Pillow>=5.3.0
PyOpenGL>=3.1.0
PyQt5>=5.11.3
pyrr>=0.9.2
scipy>=1.1.0